ebl  Use intptr_t instead of long to be sure to handle 64bit pointer
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@8389 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/lib/message.c
@@ -370,10 +370,10 @@ static void make_unique_mail_filename(JCR *jcr, POOLMEM *&name, DEST *d)
 {
    if (jcr) {
       Mmsg(name, "%s/%s.%s.%d.mail", working_directory, my_name,
-                 jcr->Job, (int)(long)d);
+                 jcr->Job, (int)(intptr_t)d);
    } else {
       Mmsg(name, "%s/%s.%s.%d.mail", working_directory, my_name,
-                 my_name, (int)(long)d);
+                 my_name, (int)(intptr_t)d);
    }
    Dmsg1(850, "mailname=%s\n", name);
 }
